pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aptx"
version = "1.0.0"
description = "APT-X: Enterprise-grade Automated Penetration Testing Framework"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "APT-X Team", email = "security@aptx.local"}
]
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Information Technology",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
]
keywords = ["pentesting", "security", "vulnerability", "scanner", "automation"]

dependencies = [
    "click>=8.1.0",
    "pyyaml>=6.0",
    "sqlalchemy>=2.0",
    "aiosqlite>=0.19.0",
    "pydantic>=2.0",
    "httpx>=0.25.0",
    "aiohttp>=3.9.0",
    "python-nmap>=0.7.1",
    "lxml>=4.9.0",
    "jinja2>=3.1.0",
    "weasyprint>=60.0",
    "markdown>=3.5",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "cryptography>=41.0",
    "rich>=13.0",
    "beautifulsoup4>=4.12.0",
    "tldextract>=5.0",
    "dnspython>=2.4.0",
    "python-multipart>=0.0.6",
    "apscheduler>=3.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]
postgres = [
    "asyncpg>=0.28.0",
    "psycopg2-binary>=2.9.0",
]

[project.scripts]
aptx = "aptx.cli:main"

[project.urls]
Homepage = "https://github.com/aptx-framework/aptx"
Documentation = "https://aptx-framework.github.io/docs"
Repository = "https://github.com/aptx-framework/aptx"

[tool.setuptools.packages.find]
where = ["."]
include = ["aptx*"]

[tool.setuptools.package-data]
aptx = [
    "data/**/*",
    "ui/templates/**/*",
    "ui/static/**/*",
    "reporting/templates/**/*",
]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311', 'py312']

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "N", "B", "C4"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
